requirements.txt
adal==1.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
numpy==1.23.4
openpyxl==3.1.2
pandas==1.5.1
pika==1.3.2
protobuf==3.20.1
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.6
redis==3.5.3
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
